Remove an include of views/../textfield.h getting compiled on Mac.
This is pulling in a bunch of views headers which don't seem to be
required in this file. Noticed some bots reporting compile errors for
mac trybots due to this.

Appeared by accident in r185787 (code was only being removed from this
file in that revision).

BUG=None
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/108913005