Tighten the blacklist on VMWare on OS X
There will no such thing as blacklisting just FCM in the future, so
first we will change these configurations to disable everything (so
they'll be like WinXP should be).

Fix up a number of tests that don't work when there is no GPU
process. It's not clear how these tests work on WinXP, so add a
temporary log statement in one of these tests so we can figure
out why they are passing on WinXP (they do not pass locally when
the GPU is blacklisted).

BUG=286038

Review URL: https://codereview.chromium.org/107093006