cc: Added invalidation frequency for picture pile.
This patch keeps track of the invalidation frequency of each of
the picture pile grid cells. When the invalidation goes over a
certain threshold (0.75) and the grid cell is further than
specified amount of pixels from the viewport (512), then the
cell is not re-recorded. For a cell that was previously skipped,
either frequency dropping below the threshold or it coming closer
than 512 pixels will cause it to re-record again.

BUG=254320

Review URL: https://codereview.chromium.org/106393003