Fix for a browser window incorrectly activating itself when we select a menu item by touch on Windows.
This scenario occurs when the wrench menu for e.g extends beyond its browser window and displays over another browser window
on the same thread. If we select any option which is outside the bounds of the menu's browser window then the item gets selected
However a WM_MOUSEACTIVATE message is sent to the other browser window causing it to get activated.

This only happens if the windows are on the same thread and is unclear as to why this happens. Appears to be a bug in the touch
infrastructure in Windows.

To workaround this we set a property on the window at the current cursor location in the MenuController code if the item
was selected by touch. We check for this property in the OnMouseActivate handler and don't activate the window if this is set.

Fixes bug https://code.google.com/p/chromium/issues/detail?id=326128

BUG=326128
R=sky@chromium.org, sky

Review URL: https://codereview.chromium.org/100193004