cc: Make BrowserCompositorInvalidateLayerTreePerfTest change mailbox.
The BrowserCompositorInvalidateLayerTreePerfTest test is meant to
simulate a renderer scrolling (ie changing its contents constantly)
from the perspective of the browser, by updating a texture mailbox
continuously. However, it seems to not be doing so any more.

This has the test contiunously SetNeedsCommit() and changes the
mailbox on each commit as intended.

Depends on https://codereview.chromium.org/107243002/

R=clholgat@chromium.org, enne, powei@chromium.org
BUG=

Review URL: https://codereview.chromium.org/107523003