Uses empty origin URL for extension crash notifications to avoid the conflict.
ShowBalloon was called asynchronously, but the comment says this is
just because the notifications will be canceled at EXTENSION_UNLOADED
event. Asynchronous call of ShowBalloon is not good, because its profile
may be destructed before it's called.

Actually this is because the crash notification has the origin of
the crashed extension. Instead the crash notification should not have
that origin, because it's not created from that extension, it's
created from the system.

When all platforms supports message center notifications, it should
also support NotifierId to note the crash notification is created
from the system.

BUG=322320
R=dewittj@chromium.org, atwilson@chromium.org
TEST=manually

Review URL: https://codereview.chromium.org/98163004