IndexedDB: Run abort tasks if backing store commit fails
If a request triggers an abort, the transaction runs its abort tasks
which restore the in-memory metadata. If the transaction commit
fails this was not done, leaving the in-memory and on-disk state
out of sync.

Ensure that the tasks are executed in this case.

BUG=241843
R=dgrogan@chromium.org

Review URL: https://codereview.chromium.org/104873003