Init LatencyInfo for TouchEvent only when constructing from NativeEvent
I am guessing that manually constructing TouchEvent with provided timestamp
can result in a wrong LatencyInfo that can contribute a huge amount of wrong
UMA data.

So lets not init the LatencyInfo if the TouchEvent is not constructed from
NateiveEvent and see if we can bring the UMA stats of
Event.Latency.RendererImpl.GestureScrollUpdate back to be reasonable.

BUG=326954
TEST=Event.Latency.RendererImpl.GestureScrollUpdate still shows when testing
     scroll on Pixel.

Review URL: https://codereview.chromium.org/99013006