#including the correct header for ssize_t.
AEAD-related code could not be compiled on Mac and Windows
because the proper header file was not included and
SSIZE_MAX macro was missing.
This patch is a sort of workaround to make them compiled on Mac and Windows.

Non goal: care about all platforms that does not have ssize_t.
See the code review for details.

Review URL: https://codereview.chromium.org/98403012