Refresh for the Chrome notifications image template.
Previously we auto-scaled all images up to fit a 360x540 rectangle.
Since this is too big, we've reduced the rectangle to 360x240.  Images
that do not match this 3:2 aspect ratio will have a border applied,
further reducing the max size to 340x220. This design allows for full
bleed images for well-behaved notifiers but also a sleek look for other
sizes of image.

BUG=326224

Review URL: https://codereview.chromium.org/102473005