Implement WebAutofillClient::openTextDataListChooser().
This function is called when a triangle mark in <input
list=...> is clicked, and it should show <datalist>
suggestions.
Unlike the normal autofill process, we'd like to skip prefix
matching for ease to replace the existing value without
keyboard operation, and make the behavior similar to <select> menu.

BUG=152375

Review URL: https://codereview.chromium.org/107353005