Add --disable-avfoundation to the command line for browser_tests.
AVfoundation is present in Mac OS X >= 10.7 (Lion).

Browser_tests in Mac uses ManyCam as a camera device. Manycam
does not work with AVFoundation, so temporarily we force QTKit
for the tests to run. If ManyCam starts being correctly 
supported in Mac with AVFoundation, this patch can be 
reverted (see bug).

(FI, Longer term we'd like to move most of the tests to 
FileVCD, and leave a very reduced amount of physical devices 
to test real hardware/OS interactions.)

BUG=327618

Review URL: https://codereview.chromium.org/112773003