Use the Android Download Manager for data reduction proxy downloads.
Downloads from sites or via proxies that require HTTP authorization
are in general downloaded by the network stack because the ADM
can't authorize. This change makes an exception. If the download is
via the data reduction proxy and the download's site does not
require proxy authorization, then use the ADM.

BUG=326225

Review URL: https://codereview.chromium.org/102953006