Pepper: Send dev/canary channel status to plugins.
As part of supporting "dev channel" methods in Pepper APIs, the plugin needs to
know whether or not it's running in a channel that supports those methods.

This patch plumbs that information through LoadPlugin so it's available by the
time get_interface<>() would ever be called.

TBR=piman
BUG=325403

Review URL: https://codereview.chromium.org/104673003