Use std::vector instead of base::List for log messages in GpuDataManagerImplPrivate
This is to verify if the crashes in ~GpuDataManagerImplPrivate() in chromeos is indeep a racing in log message handling.

If after this CL, there are no new crashes in ~GpuDataManagerImplPrivate() that are related with log_messages_, then the bug is in DictionaryValue() instead.

BUG=327463
TEST=
B=kbr@chromium.org
R=kbr@chromium.org

Review URL: https://codereview.chromium.org/101513006