Move remoting manifestdependency to external .manifest file
I plan (https://codereview.chromium.org/111373002/) to make manifest
handling simpler in gyp. In particular, doing the manifest embedding
in one pass to avoid the need to invoke the linker twice.

In order to do this /manifestdependency on the linker command line needs
to be removed as it creates dependencies that are only made clear by
having the linker generate a manifest. In this case it's simple enough
to simply use a .manifest file with the same dependency information that
can be merged beforehand.

So, move from .gyp linker flags to a file in AdditionalManifestFiles.

R=sergeyu@chromium.org
BUG=324863

Review URL: https://codereview.chromium.org/108303004