Work before enabling software compositing on Mac
Implement RenderWidgetHostViewMac::CopyFromCompositingSurface for the
software compositing path and disable copying to video frames and frame
subscription (this path needs to be functional, but not particularly
performant).

Fix a bug where RenderWidgetHostViewMac::SoftwareFrameWasFreed would
reach in to a destroyed RenderWidgetHostImpl.

Fix a bug where RenderWidgetHostViewMac::software_latency_info_ was
not updated in RenderWidgetHostViewMac::OnSwapCompositorFrame.

Fix a bug where the ack for a software frame is sent immediately, instead
of when the frame is actually drawn.

BUG=286038
R=hclam@chromium.org

Review URL: https://codereview.chromium.org/113603002