[sync] Clean up ProfileSyncServiceHarness: Eliminate state change machine
ProfileSyncServiceHarness is a class used by the desktop integration
tests to wait for adequate periods of time for various sync operations
to complete. Internally, it implements a state change machine, that over
time, has accumulated several new states, making the class brittle and
prone to bugs.

This CL is the first in a series of CLs to clean up the class, make it
less prone to errors, and as far as possible, eliminate dependencies on
internal details of implementation in ProfileSyncService.

This CL does the following:
1) Gets rid of the state change machine and WaitStates in PSSHarness.
2) Defines StatusChangeChecker, a callback mechanism that allows us to
   wait on arbitrary changes to the state of the sync engine.
3) Implements AwaitStatusChange, a function that takes a
   StatusChangeChecker and waits until its criteria are satisfied.
4) Splits up the logic in RunStateChangeMachine function into separate
   helper functions that do not depend on past WaitStates.
5) Fixes a TODO and changes EnableEncryptionForType to EnableEncryption.
6) Gets rid of the "reason" string that was being passed in by numerous
   tests while calling various Await* methods. If a call to Await* times
   out, PSSHarness will take care of logging a meaningful reason for the
   failure.
7) Fixes a bug where we were neglecting to mark sync setup as complete
   after entering a custom passphrase. (This only affects select tests
   involving custom passphrases.)

BUG=323380,95619
TEST=sync_integration_tests

Review URL: https://codereview.chromium.org/83363006