Fix potential flake in PrerenderBrowserTest.PrerenderPageNewTab.
When the swap starts, about:blank hasn't finished loading yet. If the merge
takes too long, the ProvisionalChangeToMainFrameUrl from the previous
navigation may creep in and we incorrectly cancel the pending swap. This race
isn't a problem for correctness since it is always safe to cancel the pending
swap, though it is a case where we don't swap when we potentially could in the
event of a very very tight race occurring. To avoid it flaking the test, wait
for about:blank in the new tab to load first.

Triggered flake by increasing the merge timeout and putting a sleep in
DOMStorageNamespace::Merge.

BUG=none
TEST=PrerenderBrowserTest.PrerenderPageNewTab

Review URL: https://codereview.chromium.org/115263003