Set wheel ticks for synthetic mouse wheel events.
This patch sets the wheelTicksX/Y values for synthetic mouse wheel events to the
sign of the deltaX/Y values. This is consistent with the behavior of actual
scroll events (tested on Linux). It is also required to drive pages that have a
"mousewheel" handler and query the wheelDelta(X/Y) values of the event. These
values are set using the wheelTicks values (multiplied by a constant) rather
than the delta values.

BUG=328369

Review URL: https://codereview.chromium.org/103173003