Add support for normal Git patches to path filter script.
Turns out that Git patches that are directly uploaded to a try server's
SVN repo is not processed in the same way as patches uploaded to Rietveld.
This means the implementation in https://codereview.chromium.org/27575002/
does not solve my problem with filtering try job patches for WebRTC
(since we don't have Rietveld integration for our try server).

This CL should resolve this and handle both types of patches.

BUG=312827
TEST=local tests

Review URL: https://codereview.chromium.org/113783005