Fix default start position for synthetic scroll.
If no starting coordinates are specified for synthetic scroll, we want to start
in the center of the page. The current implementation computes these coordinates
wrt to the top-left corner of the *screen*, whereas the coordinates of the input
events contain coordinates wrt to the top-left corner of the *window*. This CL
fixes that.

BUG=328422

Review URL: https://codereview.chromium.org/103053007