Files.app: Fix a bug of changing directory.
Changing directory is asynchronous step, but previously it does not care about
the previous requests to change directory. Thus the following code path is
possible.

1. Request change directory to A.
2. Request change directory to B.
3. Request of B is done.
4. Request of A is done.

This causes the flakiness of browser_tests.  This CL adds a sequential number to
check the order of requests and make only the latest request be accepted.

BUG=328775
TEST=browser_tests

Review URL: https://codereview.chromium.org/104663008