Don't silently update cached accessible bounding boxes.
We shouldn't be updating the cached bounding box of accessible
objects that haven't otherwise changed, because that prevents
other code from detecting that the location has changed and
communicating that to assistive technology.

By removing this line of code, AppendLocationChangeEvents()
is able to take care of any location changes correctly.

BUG=328517
R=aboxhall@chromium.org

Review URL: https://codereview.chromium.org/101763004