linux_aura: Lookup theme colors on GtkCustomMenu{,Item}.
In the GTK+ version of chrome, we created two custom menu classes:
GtkCustomMenu and GtkCustomMenuItem, which we used to implement the
buttons in menu concept used in the hotdog menu. Some GTK+ themes appear
to rely on these class names; GNOME 3's Adwaita is reportedly broken and
styles chrome menu's separately (and in a way that I suspect will be
fixed by this patch.)

This patch creates dummy implementations of both of these classes.

BUG=318285
R=thestig@chromium.org

Review URL: https://codereview.chromium.org/115303004