Automatically trigger installation of high-quality speech synthesis extension.
When a certain threshold number of utterances are spoken in a given session
with the built-in speech synthesis, set a preference that enables
automatic download and (silent) installation of a component extension that
provides higher-quality speech synthesis.

This change also gets rid of the code that loaded the built-in speech
synthesis component extension on-demand and instead installs it when
loading all component extensions. It now uses an event page, so the
cost of loading it is minimal.

BUG=314718
R=asargent@chromium.org, dpolukhin@chromium.org, thestig@chromium.org

Review URL: https://codereview.chromium.org/102263005