Fix linux/gcc compiler warnings in tools/ipc_fuzzer.
It's really sad when your ASCII art breaks the compilation, but ending
a //-style comment line with a \ means the comment continues and this is
flagged as a bad thing (tm).

The other change is a signed/unsigned comparison, which is OK because it
is range checked prior to casting.

BUG=

Review URL: https://codereview.chromium.org/115673002