WinAura: don't create tooltip on hide
The tooltip code is executed via timers from
TooltipController::UpdateIfRequired()

Which say a browser startup with no user intervention creates the tooltip
window only for the purposes of hiding it.

This fires in Ash on Windows creating a tooltip window in the wrong
side, it is hidden but it is something we want to avoid.

BUG=none
TBR=sadrul

Review URL: https://codereview.chromium.org/103143005