[rAC, OSX] Ensure proper navigation targets.
The rAC dialog embeds a WebContentsView to display the sign-in view.
All navigations in that sign-in view should happen according to a
specific set of rules.

1) Cross-domain navigation goes into a new tab.
2) Anything with target="blank_" including same-domain is opened as a
   new tab in the browser.
3) Same-domain navigation via link stays in the dialog's contents.
4) Same-domain POSTs stay POSTs

Item #1 is achieved by letting the delegate handle non-local toplevel
requests. This triggers OpenTabFromURL, where the code navigates on the
original page's content, instead of the dialog's content.

Item #2 is handled by AddNewContent - any time new WebContents are
created, they are routed through there and then added to the Browser
associated with the original page.

Item #3 is handled in two ways - if the RenderView considers this
non-local (i.e. there is no opener), it routes it through AddNewTab*,
which loads that URL on the dialogs WebContents. If it's considered
local, Blink handles the necessary load and it stays within the
WebContents too.

Item #4 is handled by Blink, since all form-related requests are
considered non-local.

BUG=326882
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/114503003