Always invalidate the audio device cache in the MediaStreamManager on Android.
There's no SystemMonitor on Android that will notify us when devices
are added or removed, so we need to populate the cache on every request.
There is already an up-to-date cache in the browser side audio manager
that we can rely on so the performance impact is minimal.

BUG=324464
TEST=Calling MediaStreamTrack.getSources in Chrome on Andriod should work correctly after changing the device configuration (e.g. unplug a headset).

Review URL: https://codereview.chromium.org/115823002