Guest renderers should not host non-guest content.
The problem was that guest renderers eagerly answered
"yes, I can host anything!" and this created a problem
when we were out of processes and looking for a renderer
to host a newly started tab / a new navigation.

This makes the "reload a WebView" case *not* go though
IsSuitableHost. Creating or navigating a WebView already
didn't go via IsSuitableHost before this CL.

BUG=325378

Review URL: https://codereview.chromium.org/113173006