[GCM] Reland: Delay running register/send operations when loading is in progress
When GCMProfileService and GCMClient are loading from persistent stores,
we need to delay register/send operations until the loading is all done.

A new inner class DelayedTaskController has been added to GCMProfileService.

Also, GCMClient interface has been updated to add a new method SetDelegate
that should be called to passs the delegate for a given user. We used to
rely on passing delegate in CheckIn. But it is not always called because
the check-in info could be loaded directly from the prefs store.

BUG=284553
TEST=existing tests plus new tests added

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=239676

R=fgorski@chromium.org, zea@chromium.org

Review URL: https://codereview.chromium.org/103933004