Possible fix for use of Window after deleted
An ActivationChangeObserver is trying to use the window supplied to
OnWindowActivated. It seems as though the window is not valid; most
likely deleted. I'm guessing an earlier ActivationChangeObserver
deleted it.

This change makes it so that if an ActivationChangeObserver deletes a
window we supply NULL to future ActivationChangeObservers.

BUG=239169
TEST=covered by test
R=ben@chromium.org

Review URL: https://codereview.chromium.org/115463004