Cast: Changing log classes to support incoming time of event.
Prior to this cl the log class retrived the time from the system
with this change the time of a log event can be specified by the
caller. This change is needed for the receiver of a RTCP message
that contain log messages to insert them with a correct time.

Review URL: https://codereview.chromium.org/111663003