Work around Mac OS 10.9 window server synchronization bug.
When using the integrated GPU with multiple windows open, the display
will shudder and flicker. It appears that the front and back buffers
are fighting to be displayed (sometimes the previous frame is shown,
sometimes the new frame is shown). When a glFinish is added here, the
bug goes away.

BUG=318877

Review URL: https://codereview.chromium.org/103943011