Making sandbox error message on moving to namespaces more accurate.
When trying to move to new namespaces if a feature is supported but doesn't
work we fail with an error message.

The old error message was always "Failed to move to new PID namespace" but the
cause could be a failure to move to a new "network namespace".

BUG=326039

Review URL: https://codereview.chromium.org/102723004