Fix pointer lock tests for Chrome Apps.
The focus was not guaranteed by using the ui_test_uitls call.
It was made obvious by cl 108003007 (the cl made the document no longer
focused if the window is not focused which was the case here and as a
consequence, the test was failing).

BUG=223447

Review URL: https://codereview.chromium.org/113343005