Add Android Release and Clang builders for WebRTC
Previously, we only had an 'android_ndk' trybot
(building Debug) at the tryserver. This CL renames
it to 'android' and adds a Release bot as well.

It also adds a Clang builder and trybot.

Some parts of this is already covered in
https://codereview.chromium.org/116383002/
but as those machines might take a while to get,
I'd like to land this before the holidays.

TEST=Locally firing off the Android Clang (dbg) builder and verified it used Clang to build.
BUG=webrtc:2735

Review URL: https://codereview.chromium.org/117333003