Start device policy core if policy is present.
Previously, we'd only start if the policy indicated that the client is
actively managed. That's wrong because a device in state UNMANAGED
needs to be able to fetch policy to get out of that state again. Also
add a new unit test to cover this scenario.

BUG=chromium:329203

Review URL: https://codereview.chromium.org/98143009