Ensure that we don't start the browser shutdown sequence when the last browser window in one environment (desktop or ASH) goes away as on Windows 8 we could have both running simultaneously.
The current code in Browser::OnWindowClosing was checking the current browser list size to determine if the shutdown sequence
needs to start. It should be using the chrome::ShouldStartShutdown function in application_lifetime.cc which already has
the logic to check if the other environment is valid or not.

Fixes bug http://crbug.com/313371

BUG=313371
R=cpu@chromium.org, msw@chromium.org, cpu

Review URL: https://codereview.chromium.org/118063002