[Mac] Add a rect check to determine when to animate extension popup window size changes.
Something about the new child window animation changed in 10.9. It's no longer
possible to detect a child window that's animating in based on its alphaValue,
but comparing [window frame] to [[window animator] frame] seems to work.

r240455 was a previous attempt at fixing this problem. Its policy was to
always animate on size changes, but wound up sometimes not displaying a popup
at all (bug 328776). I was able to reproduce that problem in a release build.
This version does not suffer from that defect. The new animation detection
should work on 10.9 and earlier OS releases.

BUG=307912
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/102113003