Optimize ExtensionAPI::IsAnyFeatureAvailableToContext from O(#features) to O(#children + log #features).
This also replaces O(N) string allocations with 1 and moves the top-level
feature check to the top of the function to provide a quick exit when the
context has full access.

And it adds an IsAvailable() overload taking a Feature to avoid the string
copies in GetFeatureDependency().

BUG=305000

Review URL: https://codereview.chromium.org/107473004