Race condition in component_updater::BackgroundDownloader.
After the download completes and the task is posted to the UI thread to make the download available to CUS, the downloader can't run anymore code on the FILE thread. In the buggy case, it tried to enum and clean up old jobs. The code is preempted,
the object is destroyed from the UI thread, and the BITS code crashed.


BUG=329734

Review URL: https://codereview.chromium.org/105633009