Prevent ONC-pushed certificates from being used with multiprofiles.
If a profile uses ONC-pushed certificates then it can't be part of a
multiprofile session for security reasons. This change enforces that:

- if ONC-pushed certificates are in use then multiprofiles is disabled;

- if a multiprofile session is ongoing then updates to the certificates
  pushed by ONC are ignored until the next restart;

- a profile that already used an ONC-pushed certificate before can not
  join a multiprofile session.

The last point required moving the "tainted" signal from the profile's
prefs into local_state, so that it can be checked before the profile
is loaded and added to the session.

BUG=323854

Review URL: https://codereview.chromium.org/117263002