[gin] fix casting WrappableBase to a derived class
We first need to cast the void pointer to WrappableBase before we
can cast it to the derived class: initially a WrappableBase* is
stored as holder, so we can't directly cast it to the derived class.

To test this, I added a useless base class so the test object this
pointers don't align with the WrappableBase this pointer.

Also update the comments about how to use WrappableBase.

BUG=none
R=aa@chromium.org,dcarney@chromium.org

Review URL: https://codereview.chromium.org/119173002