Added line to print version information when compiling with xcodebuild
I think we talked about this early on, but I can't find any bug filed for it so I think it got lost somewhere along the line. I may have thought it'd be easy to add with recipes, but with so many delays in converting everything to recipes, we don't have it yet on any of the builders.

BUG=329677
R=lliabraa@chromium.org, petermayo@chromium.org

Review URL: https://codereview.chromium.org/100993006