Set the active input method correctly when EnableLayouts is called.
ChangeInputMethodInternal called by EnableLayouts checks if component_extension_ime_manager_ is initialized or not.  When it is not initialized, active_input_method is not updated even if the new input method is a keyboard layout.

This patch checks if the new input method is a keyboard layout or not too.

BUG=329061

Review URL: https://codereview.chromium.org/118273002