Reset is_completed_ before attempting a new crx component download.
When downloading a url, the state of the BackgroundDownloader must be
reset.

A better solution later on would be implementing a BitsFetcher that can
only do one download, similar to UrlFetcher. This avoids the situation
when dirty state post-download is carried over to a subsequent
download.

BUG=304354

Review URL: https://codereview.chromium.org/102993014