[Telemetry] Don't restore command lines that will break Chrome.
A common problem is that if something goes wrong with Telemetry, the command
line file which points all network traffic at Web Page Replay can be left in
place which results in users not understanding why no pages will load.

So if we detect this setting, simply clear the file.

BUG=

Review URL: https://codereview.chromium.org/108573008