Reland "Add a HistogramRecorder class and use cases."
This CL adds a utility class to streamline the task of testing that metrics have been updated as expected. Specifically, the HistogramRecorder class allows a client to obtain the differential value of a given histogram in the interval since the given HistogramRecorder instance was created.

This reverts commit b957026f43ef9464a8cdd5a240867c807b2281da.

BUG=232414
TBR=rsleevi, mark, groby, thakis

Review URL: https://codereview.chromium.org/19866004