Restart fullscreen video playback when switching back from background
This cl is supposed to improve user experience. The idea is that it
is likely that the intention of the user is to continue video playback
when switching back to the browser.  Additionally this unifies the behavior
of switching back from background and from device lock. The new behavior
matches with Safari on IOS and Chrome on desktop.

Now we destroy the fullscreen view when the application is stopped and
recreate it when the application comes back to foreground.

Bug=323697

Review URL: https://codereview.chromium.org/103583005