Cleanup storage_partition_impl_unittest.cc.
Removes:
 - unnecessary casts
 - unnecessary conversions to pointers
 - unnecessary wrapper over RunLoop.

It also removes stack-based creation of the StoragePartition and temp
directories. This was fragile and unnecessary. The TestBrowserContext itself
already created a default one.  The stack-based version was actually coopting
contexts that were created in the default partition to create a partial
stub. This unfortunately also avoided most of the initialization code and
checks in the StoragePartitionImpl::Create() function leading to more
fragility.

Now, each test case only has one default StoragePartition which is correctly
tied to the containing BrowserContext. The contexts in this partition are
also no longer shared by the stubs.

TBR=boliu
BUG=none

Review URL: https://codereview.chromium.org/101423007