remove plugin exception restoration code
First part (this CL): remove only caller.

Back in the days of Flash 9, we added this code to restore
the unhandled exception handler which flash sometimes stole.

Since flash 10 this is no longer true, and now it is not possible
in chrome to use old flashes so this code just adds complexity
for no good reason.

TBR=jam
BUG=none

Review URL: https://codereview.chromium.org/122553003