cc: Improve worker pool performance by using std:make_heap instead of std:priority_queue.
This reduces the complexity of building a new "ready to run" queue in
WorkerPool::SetTaskGraph() from O(N·log(N)) to O(N).

This refactor is also useful for supporting shared worker threads as it
makes it easy to reuse the compare function and maintain a heap of
ready to run task namespaces.

BUG=246546

Review URL: https://codereview.chromium.org/123113002