Recreate the compositor when we receive the target HWND from the metro viewer process on Windows 8 ASH.
This ensures that the compositor paints on the correct surface.

Fixes a regression which occurred after the latest aura refactoring changes. This resulted in nothing painting
in Windows 8 ASH.

Ben's change to address the metro painting issue was reverted on Dec 20th due to ash_unittest crashes.

This fix works well with the ash_unittests and regular Chrome code path.

BUG=330179
R=cpu@chromium.org, cpu
TBR=beng

Review URL: https://codereview.chromium.org/123563003