Only close chrome://settings/clearBrowserData if it is visible
The chrome://settings/clearBrowserData overlay is closed when the clear
operation has finished. This CL ensures that we only close the overlay if it is
actually visible. If the user navigated to a different overlay, that overlay
should not be closed when the browsing data has been cleared.

BUG=330706
R=dbeam@chromium.org

Review URL: https://codereview.chromium.org/114863006